# docs/numerology.openapi.yaml
openapi: 3.1.0
info:
  title: Numora API - Numerology
  version: 1.0.0
  description: >
    OpenAPI contract for numerology calculation endpoint.
servers:
  - url: http://localhost:3000
paths:
  /api/readings:
    get:
      summary: List authenticated user readings
      operationId: listReadings
      tags:
        - Readings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Readings fetched successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - readings
                properties:
                  readings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Reading"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a reading for authenticated user
      operationId: createReading
      tags:
        - Readings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReadingInput"
            example:
              firstName: Marie
              lastName: Dupont
              birthDate: 1990-03-15
              category: life-path
              results:
                lifePath: 1
                expression: 7
      responses:
        "201":
          description: Reading created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - reading
                properties:
                  reading:
                    $ref: "#/components/schemas/Reading"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/readings/{id}:
    get:
      summary: Get one reading by id for authenticated user
      operationId: getReadingById
      tags:
        - Readings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reading fetched successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - reading
                properties:
                  reading:
                    $ref: "#/components/schemas/Reading"
        "400":
          description: Invalid reading id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Reading not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete one reading by id for authenticated user
      operationId: deleteReadingById
      tags:
        - Readings
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Reading deleted successfully
        "400":
          description: Invalid reading id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Reading not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/numerology/calculate:
    post:
      summary: Calculate full numerology profile
      operationId: calculateNumerology
      tags:
        - Numerology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NumerologyCalculationInput"
            examples:
              basic:
                summary: Minimal payload
                value:
                  fullName: Marie Dupont
                  birthDate: 1990-03-15
              complete:
                summary: Full payload
                value:
                  fullName: Marie Dupont
                  birthDate: 1990-03-15
                  referenceDate: 2026-04-14
                  address:
                    streetNumber: "12"
                    streetName: Rue de Paris
                    allowMasterNumbers: true
                  locality:
                    postalCode: "75001"
                    city: Paris
                    allowMasterNumbers: true
      responses:
        "200":
          description: Numerology profile calculated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - result
                properties:
                  result:
                    $ref: "#/components/schemas/NumerologyResult"
        "400":
          description: Validation or calculation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Reading:
      type: object
      required:
        - id
        - userId
        - firstName
        - lastName
        - birthDate
        - category
        - results
        - createdAt
      properties:
        id:
          type: string
        userId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        category:
          type: string
          enum:
            - life-path
            - compatibility
            - forecast
            - custom
        results:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    CreateReadingInput:
      type: object
      required:
        - firstName
        - lastName
        - birthDate
        - category
        - results
      properties:
        firstName:
          type: string
          minLength: 1
          example: Marie
        lastName:
          type: string
          minLength: 1
          example: Dupont
        birthDate:
          type: string
          format: date
          example: 1990-03-15
        category:
          type: string
          enum:
            - life-path
            - compatibility
            - forecast
            - custom
        results:
          type: object
          additionalProperties: true
    NumerologyCalculationInput:
      type: object
      required:
        - fullName
        - birthDate
      properties:
        fullName:
          type: string
          minLength: 1
          example: Marie Dupont
        birthDate:
          type: string
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
          example: 1990-03-15
        referenceDate:
          type: string
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
          example: 2026-04-14
        address:
          $ref: "#/components/schemas/AddressInput"
        locality:
          $ref: "#/components/schemas/LocalityInput"
    AddressInput:
      type: object
      required:
        - streetNumber
        - streetName
      properties:
        streetNumber:
          type: string
          example: "12"
        streetName:
          type: string
          example: Rue de Paris
        allowMasterNumbers:
          type: boolean
          default: true
    LocalityInput:
      type: object
      required:
        - postalCode
        - city
      properties:
        postalCode:
          type: string
          example: "75001"
        city:
          type: string
          example: Paris
        allowMasterNumbers:
          type: boolean
          default: true
    NumerologyResult:
      type: object
      required:
        - identity
        - core
        - personal
        - challenges
        - karmic
        - matrixDestiny
        - treeOfLife
        - universalVibrations
        - business
        - place
      properties:
        identity:
          type: object
          required:
            - fullName
            - birthDate
          properties:
            fullName:
              type: string
            birthDate:
              type: string
        core:
          type: object
          required:
            - lifePath
            - expression
            - soulUrge
            - personality
            - birthday
            - heart
            - realisation
          properties:
            lifePath: { type: integer }
            expression: { type: integer }
            soulUrge: { type: integer }
            personality: { type: integer }
            birthday: { type: integer }
            heart: { type: integer }
            realisation: { type: integer }
        personal:
          $ref: "#/components/schemas/PersonalNumbersResult"
        challenges:
          type: object
          required:
            - challengeNumbers
            - lifeCycles
            - realizationPeriods
          properties:
            challengeNumbers:
              type: object
              additionalProperties:
                type: object
                properties:
                  number: { type: integer }
                  description: { type: string }
            lifeCycles:
              type: object
              properties:
                firstCycle: { type: integer }
                secondCycle: { type: integer }
                thirdCycle: { type: integer }
            realizationPeriods:
              type: object
              properties:
                firstPeriod: { type: integer }
                secondPeriod: { type: integer }
                thirdPeriod: { type: integer }
                fourthPeriod: { type: integer }
        karmic:
          type: object
          properties:
            karmicNumbers:
              type: object
              additionalProperties: true
            cycleKarmicNumbers:
              type: object
              additionalProperties: true
            karmicDebts:
              type: object
              additionalProperties: true
        matrixDestiny:
          type: object
          additionalProperties: true
        treeOfLife:
          type: object
          additionalProperties: true
        universalVibrations:
          type: object
          required:
            - day
            - month
            - year
            - universal
          properties:
            day: { type: integer }
            month: { type: integer }
            year: { type: integer }
            universal: { type: integer }
        business:
          $ref: "#/components/schemas/BusinessNumbersResult"
        place:
          type: object
          properties:
            addressVibration:
              type: integer
            localityVibration:
              type: integer
    PersonalNumbersResult:
      type: object
      required:
        - year
        - month
        - day
      properties:
        year:
          $ref: "#/components/schemas/PersonalNumberItem"
        month:
          $ref: "#/components/schemas/PersonalNumberItem"
        day:
          $ref: "#/components/schemas/PersonalNumberItem"
    PersonalNumberItem:
      type: object
      required:
        - number
        - description
      properties:
        number:
          type: integer
        description:
          type: string
    BusinessNumbersResult:
      type: object
      required:
        - expression
        - active
        - hereditary
      properties:
        expression:
          $ref: "#/components/schemas/BusinessNumberItem"
        active:
          $ref: "#/components/schemas/BusinessNumberItem"
        hereditary:
          $ref: "#/components/schemas/BusinessNumberItem"
    BusinessNumberItem:
      type: object
      required:
        - raw
        - value
      properties:
        raw:
          type: integer
        value:
          type: integer
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string